requirements.txt
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.7.0
jdcal==1.4.1
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
xlrd==1.2.0
zipp==3.1.0
